working program and test
