+++ src/sources/currencyLayer.js
@@ -1,36 +1,13 @@
  import config from '../config'
  import extract from '../util/extract.js'
  import load from '../util/load'
-
+ import transform from '../transformer/currencyLayer'
  const props = {
    isProduction: config.get('env') === 'production',
    file: './stubs/currencyLayer.json',
    sourceUrl:`http://apilayer.net/api/live?access_key=${config.get('currencyLayer.token')}0&currencies=USD,AUD,CAD,PLN,MXN&format=1`,
    apiUrl: config.get('api.url'),
    currencies: ['USDPLN', 'USDCAD', 'USDMXN']
- }
-
- const transform = (props, response) => {
-   const data = JSON.parse(response)
-   if (!data || !data.success) {
-     throw('Fixer return invalid response')
-   }
-
-   const { quotes } = data
-   if (!Object.keys(quotes)) {
-     throw('Fixer return invalid response')
-   }
-
-   const wantedRates = Object.keys(quotes).reduce((accumulator, currency) => {
-     if (props.currencies.includes(currency)) {
-       accumulator.push({
-         [currency]: quotes[currency]
-       })
-     }
-     return accumulator
-   }, [])
-
-   return wantedRates
  }
 
  const main = async () => {
@@ -41,3 +18,5 @@
  }
 
  export default main
+
+
